Making ContentList a component to be displayed on the home page rather than a page itself
